+++ cli/src/command/wallet.rs
@@ -8,6 +8,7 @@ use iota_sdk::{
     client::{
         constants::SHIMMER_COIN_TYPE,
         secret::{stronghold::StrongholdSecretManager, SecretManager},
+        stronghold::StrongholdAdapter,
     },
     wallet::{ClientOptions, Wallet},
 };
@@ -53,6 +54,11 @@ pub enum WalletCommand {
     ChangePassword,
     /// Initialize the wallet.
     Init(InitParameters),
+    /// Migrate a stronghold snapshot v2 to v3.
+    MigrateStrongholdSnapshotV2ToV3 {
+        /// Path of the to be migrated stronghold file. "./stardust-cli-wallet.stronghold" if nothing provided.
+        path: Option<String>,
+    },
     /// Generate a random mnemonic.
     Mnemonic,
     /// Create a new account.
@@ -148,6 +154,21 @@ pub async fn init_command(
     }
 
     Ok(wallet)
+}
+
+pub async fn migrate_stronghold_snapshot_v2_to_v3_command(path: Option<String>) -> Result<(), Error> {
+    let password = get_password("Stronghold password", false)?;
+    StrongholdAdapter::migrate_snapshot_v2_to_v3(
+        path.as_deref().unwrap_or(DEFAULT_STRONGHOLD_SNAPSHOT_PATH),
+        &password,
+        "wallet.rs",
+        100,
+        None,
+        None,
+    )?;
+    println_log_info!("Stronghold snapshot successfully migrated from v2 to v3.");
+
+    Ok(())
 }
 
 pub async fn mnemonic_command() -> Result<(), Error> {